requirements.txt
flask=1.0.2
keras=2.2.4
keras-applications=1.0.8
keras-preprocessing=1.1.0
Pillow=6.1.0
tensorflow=1.13.1
imutils=0.5.2
pickleshare=0.7.4
opencv-python-headless=4.1.0.25
numpy=1.16.4